requirements.txt
python-dotenv
discord.py
wavelink
pyyaml